Refactor : MemberValidator interface -> abstract로 변경
