+++ src/main/java/project/study/dto/login/validator/DefaultMemberValidator.java
@@ -4,7 +4,6 @@ import jakarta.servlet.http.HttpServletResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
 import project.study.domain.Freeze;
-import project.study.domain.Member;
 import project.study.dto.login.requestdto.RequestDefaultSignupDto;
 import project.study.dto.login.requestdto.RequestSignupDto;
 import project.study.dto.login.responsedto.Error;
@@ -17,40 +16,30 @@ import project.study.jpaRepository.BasicJpaRepository;
 import project.study.jpaRepository.MemberJpaRepository;
 import project.study.repository.FreezeRepository;
 
-import java.time.LocalDateTime;
-import java.util.Optional;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
 
 @Component
-@RequiredArgsConstructor
-public class DefaultMemberValidator implements MemberValidator {
+public class DefaultMemberValidator extends MemberValidator {
 
-
-    private final FreezeRepository freezeRepository;
     private final BasicJpaRepository basicJpaRepository;
     private final MemberJpaRepository memberJpaRepository;
 
-    @Override
-    public void validLogin(Member member, HttpServletResponse response) {
-        System.out.println("회원 검증 로직 시작");
-        // 이용정지 회원인지 확인
-        if (member.isExpireMember()) throw new ExpireMemberLoginException(); // 탈퇴한 회원인지 확인
-        if(!member.isFreezeMember()) return; // 이용정지된 회원이 아님
-
-        Optional<Freeze> findFreeze = freezeRepository.findByMemberId(member.getMemberId());
-        if (findFreeze.isEmpty()) return; // Freeze Entity 없음 ( 혹시 모를 예외 처리 )
-
-        Freeze freeze = findFreeze.get();
-        if (freeze.isFinish()) { // Freeze Entity는 존재하지만 이용정지 기간에 풀린 경우
-            freezeRepository.delete(freeze, member);
-            return;
-        }
-        throw new FreezeMemberLoginException(freeze.printMessage()); // 모든 조건에 걸리지 않은 회원은 이용정지 회원임.
+    public DefaultMemberValidator(FreezeRepository freezeRepository, BasicJpaRepository basicJpaRepository, MemberJpaRepository memberJpaRepository) {
+        super(freezeRepository);
+        this.basicJpaRepository = basicJpaRepository;
+        this.memberJpaRepository = memberJpaRepository;
     }
 
-
+    @Override
+    void freezeMemberLoginException(Freeze freeze, HttpServletResponse response) {
+        throw new FreezeMemberLoginException(freeze.printMessage());
+    }
+    @Override
+    void expireMemberLoginException(HttpServletResponse response) {
+        throw new ExpireMemberLoginException();
+    }
 
     @Override
     public void validSignup(RequestSignupDto signupDto) {
